Update pf_discord_rpc.cpp
Fixed console messages that incorrectly declared the date and time of compilation when it was actually just calling for the current time with the timezone hardcoded.

I did try to figure out how to inject the timezone into this as part of the compiling process, but apparently that's incredibly involved and can cause issues, so I've refrained for now.